database finished

+++ backend/db/models.py
@@ -12,6 +12,10 @@ class DataBase:
         self.RoomTable.create_table()
         self.BillTable.create_table()
         self.DetailedItemTable.create_table()
+        for i in range(10):
+            self.RoomTable.create(room_id=i + 1,
+                                  date_in=datetime.datetime.now(),
+                                  is_empty=True)
 
     class BaseModel(Model):
         class Meta:
@@ -90,7 +94,10 @@ class DataBase:
         return fee
 
     def get_bill(self, room_id):
-        date_in = self.RoomTable.filter(room_id=room_id)[0].date_in
+        have_record = self.RoomTable.filter(room_id=room_id)
+        if len(have_record) == 0:
+            return []
+        date_in = have_record[0].date_in
         bills = self.BillTable.filter(room_id=room_id)
         total_fee = 0
         for bill in bills:
@@ -100,6 +107,8 @@ class DataBase:
 
     def get_detailed_list(self, room_id):
         bills = self.BillTable.filter(room_id=room_id)
+        if len(bills) == 0:
+            return []
         ans = []
         for bill in bills:
             detailed_item = self.DetailedItemTable.filter(detailed_item_id=bill.detailed_item_id)[0]
@@ -116,8 +125,16 @@ class DataBase:
         return ans
 
     def check_in(self):
-        room_id = self.RoomTable.create(date_in=datetime.datetime.now(), is_empty=False)
-        return room_id
+        empty_rooms = self.RoomTable.filter(is_empty=True)
+        if len(empty_rooms) != 0:
+            room_id = empty_rooms[0].room_id
+            query = self.RoomTable.update(is_empty=False, date_in=datetime.datetime.now())\
+                .where(self.RoomTable.room_id == room_id)
+            query.execute()
+            return room_id
+        else:
+            new_room = self.RoomTable.create(is_empty=False, date_in=datetime.datetime.now())
+            return new_room.room_id
 
     def use_air_conditioner(self, room_id, start_temp, target_temp, mode, speed, fee_rate):
         detailed_item_id = self.DetailedItemTable.create(room_id=room_id,
@@ -138,8 +155,12 @@ class DataBase:
     def check_out(self, room_id):
         bill = self.get_bill(room_id)
         detailed_list = self.get_detailed_list(room_id)
-        query = self.RoomTable.delete().where(self.RoomTable.room_id == room_id)
-        query.execute()
+        if room_id > 10:
+            query = self.RoomTable.delete().where(self.RoomTable.room_id == room_id)
+            query.execute()
+        else:
+            query = self.RoomTable.update(is_empty=True).where(self.RoomTable.room_id == room_id)
+            query.execute()
         query = self.BillTable.delete().where(self.BillTable.room_id == room_id)
         query.execute()
         query = self.DetailedItemTable.delete().where(self.DetailedItemTable.room_id == room_id)
@@ -159,6 +180,54 @@ detail_id = db.use_air_conditioner(room, 25, 22, "cold", "mid", 1)
 sleep(10)  # 开了10s
 db.stop_air_conditioner(detail_id)  # 关掉
 b, d = db.check_out(room)  # 退房，返回账单、详单
+print(b)  # 打印账单
+print(d)  # 打印详单
+room1 = db.check_in()  # 入住，返回房间号
+room2 = db.check_in()  # 入住，返回房间号
+room3 = db.check_in()  # 入住，返回房间号
+room4 = db.check_in()  # 入住，返回房间号
+room5 = db.check_in()  # 入住，返回房间号
+room6 = db.check_in()  # 入住，返回房间号
+room7 = db.check_in()  # 入住，返回房间号
+room8 = db.check_in()  # 入住，返回房间号
+room9 = db.check_in()  # 入住，返回房间号
+room10 = db.check_in()  # 入住，返回房间号
+room11 = db.check_in()  # 入住，返回房间号
+room12 = db.check_in()  # 入住，返回房间号
+b, d = db.check_out(room1)  # 退房，返回账单、详单
+print(b)  # 打印账单
+print(d)  # 打印详单
+b, d = db.check_out(room2)  # 退房，返回账单、详单
+print(b)  # 打印账单
+print(d)  # 打印详单
+b, d = db.check_out(room3)  # 退房，返回账单、详单
+print(b)  # 打印账单
+print(d)  # 打印详单
+b, d = db.check_out(room4)  # 退房，返回账单、详单
+print(b)  # 打印账单
+print(d)  # 打印详单
+b, d = db.check_out(room5)  # 退房，返回账单、详单
+print(b)  # 打印账单
+print(d)  # 打印详单
+b, d = db.check_out(room6)  # 退房，返回账单、详单
+print(b)  # 打印账单
+print(d)  # 打印详单
+b, d = db.check_out(room7)  # 退房，返回账单、详单
+print(b)  # 打印账单
+print(d)  # 打印详单
+b, d = db.check_out(room8)  # 退房，返回账单、详单
+print(b)  # 打印账单
+print(d)  # 打印详单
+b, d = db.check_out(room9)  # 退房，返回账单、详单
+print(b)  # 打印账单
+print(d)  # 打印详单
+b, d = db.check_out(room10)  # 退房，返回账单、详单
+print(b)  # 打印账单
+print(d)  # 打印详单
+b, d = db.check_out(room11)  # 退房，返回账单、详单
+print(b)  # 打印账单
+print(d)  # 打印详单
+b, d = db.check_out(room12)  # 退房，返回账单、详单
 print(b)  # 打印账单
 print(d)  # 打印详单
